Feat : MainViewModel 통합 - #28
